// app/xmlns/www.musicxml.org.d.ts
interface BaseType {
	_exists: boolean;
	_namespace: string;
}
/** The above-below type is used to indicate whether one element appears above or below another element. */
export type aboveBelow = ("above" | "below");
// ...
interface _attributes extends BaseType {
	/** Clefs are represented by a combination of sign, line, and clef-octave-change elements. */
	clef?: clef[];
	/** Directives are like directions, but can be grouped together with attributes for convenience. This is typically used for tempo markings at the beginning of a piece of music. This element has been deprecated in Version 2.0 in favor of the directive attribute for direction elements. Language names come from ISO 639, with optional country subcodes from ISO 3166. */
	directive?: attributesDirectiveType[];
	/** Musical notation duration is commonly represented as fractions. The divisions element indicates how many divisions per quarter note are used to indicate a note's duration. For example, if duration = 1 and divisions = 2, this is an eighth note duration. Duration and divisions are used directly for generating sound output, so they must be chosen to take tuplets into account. Using a divisions element lets us use just one number to represent a duration for each note in the score, while retaining the full power of a fractional representation. If maximum compatibility with Standard MIDI 1.0 files is important, do not have the divisions value exceed 16383. */
	divisions?: number;
	footnote?: formattedText;
	/** The instruments element is only used if more than one instrument is represented in the part (e.g., oboe I and II where they play together most of the time). If absent, a value of 1 is assumed. */
	instruments?: number;
	/** The key element represents a key signature. Both traditional and non-traditional key signatures are supported. The optional number attribute refers to staff numbers. If absent, the key signature applies to all staves in the part. */
	key?: key[];
	level?: level;
	/** A measure-style indicates a special way to print partial to multiple measures within a part. This includes multiple rests over several measures, repeats of beats, single, or multiple measures, and use of slash notation. */
	measureStyle?: measureStyle[];
	/** The part-symbol element indicates how a symbol for a multi-staff part is indicated in the score. */
	partSymbol?: partSymbol;
	/** The staff-details element is used to indicate different types of staves. */
	staffDetails?: staffDetails[];
	/** The staves element is used if there is more than one staff represented in the given part (e.g., 2 staves for typical piano parts). If absent, a value of 1 is assumed. Staves are ordered from top to bottom in a part in numerical order, with staff 1 above staff 2. */
	staves?: number;
	/** Time signatures are represented by the beats element for the numerator and the beat-type element for the denominator. */
	time?: time[];
	/** If the part is being encoded for a transposing instrument in written vs. concert pitch, the transposition must be encoded in the transpose element using the transpose type. */
	transpose?: transpose[];
}
export interface attributes extends _attributes { constructor: { new(): attributes }; }
// ...
interface _barline extends BaseType {
	$coda: string;
	divisions: number;
	location: rightLeftmiddle;
	$segno: string;
	barStyle?: barStylecolor;
	coda?: emptyPrintstylealign;
	ending?: ending;
	fermata?: fermata[];
	footnote?: formattedText;
	level?: level;
	repeat?: repeat;
	segno?: emptyPrintstylealign;
	wavyLine?: wavyLine;
}
export interface barline extends _barline { constructor: { new(): barline }; }
// ...
interface _clef extends BaseType {
	additional: yesNo;
	afterBarline: yesNo;
	color: string;
	defaultX: number;
	defaultY: number;
	fontFamily: string;
	fontSize: fontSize;
	fontStyle: fontStyle;
	fontWeight: fontWeight;
	number: number;
	printObject: yesNo;
	relativeX: number;
	relativeY: number;
	size: symbolSize;
	/** The clef-octave-change element is used for transposing clefs. A treble clef for tenors would have a value of -1. */
	clefOctavechange?: number;
	/** Line numbers are counted from the bottom of the staff. Standard values are 2 for the G sign (treble clef), 4 for the F sign (bass clef), 3 for the C sign (alto clef) and 5 for TAB (on a 6-line staff). */
	line?: number;
	/** The sign element represents the clef symbol. */
	sign: clefSign;
}
export interface clef extends _clef { constructor: { new(): clef }; }
// ...
interface _direction extends BaseType {
	directive: yesNo;
	placement: aboveBelow;
	directionType: directionType[];
	footnote?: formattedText;
	level?: level;
	offset?: offset;
	sound?: sound;
	/** Staff assignment is only needed for music notated on multiple staves. Used by both notes and directions. Staff values are numbers, with 1 referring to the top-most staff in a part. */
	staff?: number;
	voice?: string;
}
export interface direction extends _direction { constructor: { new(): direction }; }
// ...
export interface directionType extends _directionType { constructor: { new(): directionType }; }
// ...
interface _emptyPlacement extends BaseType {
	color: string;
	defaultX: number;
	defaultY: number;
	fontFamily: string;
	fontSize: fontSize;
	fontStyle: fontStyle;
	fontWeight: fontWeight;
	placement: aboveBelow;
	relativeX: number;
	relativeY: number;
}
export interface emptyPlacement extends _emptyPlacement { constructor: { new(): emptyPlacement }; }
// ...
export interface formattedText extends _formattedText { constructor: { new(): formattedText }; }
// ...
interface _key extends BaseType {
	color: string;
	defaultX: number;
	defaultY: number;
	fontFamily: string;
	fontSize: fontSize;
	fontStyle: fontStyle;
	fontWeight: fontWeight;
	number: number;
	printObject: yesNo;
	relativeX: number;
	relativeY: number;
	cancel?: cancel;
	fifths: number;
	/** Non-traditional key signatures can be represented using the Humdrum/Scot concept of a list of altered tones. The key-accidental element indicates the accidental to be displayed in the key signature, represented in the same manner as the accidental element. It is used for disambiguating microtonal accidentals. */
	keyAccidental?: accidentalValue[];
	/** Non-traditional key signatures can be represented using the Humdrum/Scot concept of a list of altered tones. The key-alter element represents the alteration for a given pitch step, represented with semitones in the same manner as the alter element. */
	keyAlter?: number[];
	/** The optional list of key-octave elements is used to specify in which octave each element of the key signature appears. */
	keyOctave?: keyOctave[];
	/** Non-traditional key signatures can be represented using the Humdrum/Scot concept of a list of altered tones. The key-step element indicates the pitch step to be altered, represented using the same names as in the step element. */
	keyStep?: step[];
	mode?: string;
}
export interface key extends _key { constructor: { new(): key }; }
// ...
interface _note extends BaseType {
	attack: number;
	color: string;
	defaultX: number;
	defaultY: number;
	dynamics: number;
	endDynamics: number;
	fontFamily: string;
	fontSize: fontSize;
	fontStyle: fontStyle;
	fontWeight: fontWeight;
	pizzicato: yesNo;
	printDot: yesNo;
	printLyric: yesNo;
	printObject: yesNo;
	printSpacing: yesNo;
	relativeX: number;
	relativeY: number;
	release: number;
	timeOnly: string;
	accidental?: accidental;
	beam?: beam[];
	/** The chord element indicates that this note is an additional chord tone with the preceding note. The duration of this note can be no longer than the preceding note. In MuseData, a missing duration indicates the same length as the previous note, but the MusicXML format requires a duration for chord notes too. */
	chord?: empty[];
	/** The cue element indicates the presence of a cue note. */
	cue: empty;
	/** One dot element is used for each dot of prolongation. The placement element is used to specify whether the dot should appear above or below the staff line. It is ignored for notes that appear on a staff space. */
	dot?: emptyPlacement[];
	/** Duration is a positive number specified in division units. This is the intended duration vs. notated duration (for instance, swing eighths vs. even eighths, or differences in dotted notes in Baroque-era music). Differences in duration specific to an interpretation or performance should use the note element's attack and release attributes. */
	duration: number[];
	footnote?: formattedText;
	grace: grace;
	instrument?: instrument;
	level?: level;
	lyric?: lyric[];
	notations?: notations[];
	notehead?: notehead;
	noteheadText?: noteheadText;
	pitch: pitch[];
	play?: play;
	rest: rest[];
	/** Staff assignment is only needed for music notated on multiple staves. Used by both notes and directions. Staff values are numbers, with 1 referring to the top-most staff in a part. */
	staff?: number;
	stem?: stem;
	tie?: tie[];
	timeModification?: timeModification;
	type?: noteType;
	unpitched: unpitched[];
	voice?: string;
}
export interface note extends _note { constructor: { new(): note }; }
// ...
interface _noteType extends _noteTypevalue {
	size: symbolSize;
}
export interface noteType extends _noteType { constructor: { new(): noteType }; }
export var noteType: { new(): noteType };

/** The note-type type is used for the MusicXML type element and represents the graphic note type, from 1024th (shortest) to maxima (longest). */
export type noteTypevalue = ("1024th" | "512th" | "256th" | "128th" | "64th" | "32nd" | "16th" | "eighth" | "quarter" | "half" | "whole" | "breve" | "long" | "maxima");
// ...
interface _print extends BaseType {
	blankPage: number;
	newPage: yesNo;
	newSystem: yesNo;
	pageNumber: string;
	staffSpacing: number;
	measureLayout?: measureLayout;
	measureNumbering?: measureNumbering;
	pageLayout?: pageLayout;
	partAbbreviationdisplay?: nameDisplay;
	partNamedisplay?: nameDisplay;
	staffLayout?: staffLayout[];
	systemLayout?: systemLayout;
}
export interface print extends _print { constructor: { new(): print }; }
// ...
export interface rest extends _rest { constructor: { new(): rest }; }
// ...
interface _ScorePartwiseType extends BaseType {
	version: string;
	credit?: credit[];
	defaults?: defaults;
	identification?: identification;
	/** The movement-number element specifies the number of a movement. */
	movementNumber?: string;
	/** The movement-title element specifies the title of a movement, not including its number. */
	movementTitle?: string;
	part: ScorePartwiseTypePartType[];
	partList: partList;
	work?: work;
}
export interface ScorePartwiseType extends _ScorePartwiseType { constructor: { new(): ScorePartwiseType }; }
export var ScorePartwiseType: { new(): ScorePartwiseType};

interface _ScorePartwiseTypePartType extends BaseType {
	id: string;
	measure: ScorePartwiseTypePartTypeMeasureType[];
}
export interface ScorePartwiseTypePartType extends _ScorePartwiseTypePartType { constructor: { new(): ScorePartwiseTypePartType }; }
export var ScorePartwiseTypePartType: {new ():ScorePartwiseTypePartType};

interface _ScorePartwiseTypePartTypeMeasureType extends BaseType {
	implicit: yesNo;
	nonControlling: yesNo;
	number: string;
	width: number;
	attributes?: attributes[];
	backup?: backup[];
	barline?: barline[];
	bookmark?: bookmark[];
	direction?: direction[];
	figuredBass?: figuredBass[];
	forward?: forward[];
	grouping?: grouping[];
	harmony?: harmony[];
	link?: link[];
	note?: note[];
	print?: print[];
	sound?: sound[];
}
export interface ScorePartwiseTypePartTypeMeasureType extends _ScorePartwiseTypePartTypeMeasureType { constructor: { new(): ScorePartwiseTypePartTypeMeasureType }; }
// ...
interface ScoreTimewiseType extends _ScoreTimewiseType { constructor: { new(): ScoreTimewiseType }; }
// ...
export interface stem extends _stem { constructor: { new(): stem }; }
export var stem: { new(): stem };

/** The stem type represents the notated stem direction. */
export type stemValue = ("down" | "up" | "double" | "none");
// ...
export type yesNo = ("yes" | "no");
// ...
export interface document extends BaseType {
	/** The score-partwise element is the root element for a partwise MusicXML score. It includes a score-header group followed by a series of parts with measures inside. The document-attributes attribute group includes the version attribute. */
	scorePartwise: ScorePartwiseType;
	/** The score-timewise element is the root element for a timewise MusicXML score. It includes a score-header group followed by a series of measures with parts inside. The document-attributes attribute group includes the version attribute. */
	scoreTimewise: ScoreTimewiseType;
}